Silence signed/unsigned comparison warnings in GCC.
i2c_frequency() compares a uint32_t ref variable to the int hz
function parameter passed in by the caller.  I forced this to be an
uint32_t comparison.

i2c_slave_write() declared i and count variables to be of type uint32_t
but used them as int type throughout the code (in comparisons and
returns) so I switched them to be of signed int type.

spi_frequency() contains a change similar to that made in
i2c_frequency().

+++ libraries/mbed/targets/hal/TARGET_Freescale/TARGET_KL25Z/spi_api.c
@@ -153,7 +153,7 @@ void spi_frequency(spi_t *obj, int hz) {
         divisor = 2;
         for (spr = 0; spr <= 8; spr++, divisor *= 2) {
             ref = PCLK / (prescaler*divisor);
-            if (ref > hz)
+            if (ref > (uint32_t)hz)
                 continue;
             error = hz - ref;
             if (error < p_error) {
